Single source WebSocket and WebRTC connections in App.js
