+++ client/package.json
@@ -19,11 +19,8 @@
     "react-bootstrap": "^1.0.0-beta.12",
     "react-codemirror2": "^6.0.0",
     "react-dom": "^16.9.0",
-    "react-native": "^0.60.5",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.1.1",
-    "react-scroll": "^1.7.14",
-    "react-scroll-to-component": "^1.0.2",
     "reactstrap": "^8.0.1",
     "socket.io-client": "^2.2.0",
     "styled-components": "^4.3.2"
